Add BOM building code (#11)
* Add BOM building code

New API for the process of building the CBOM itself. The
`cyclone-dx` itself has no such API. This adds a builder pattern
allowing BOM to be gradually updated with new data. Add an initial
header and a metadata.

* produce CBOM from x509 certificate

Add code generating a cdx.Component from an *x509.Certificate and use a
bom.Builder to build the results. The binary now prints the CBOM JSON to
stdout.

* produce UTC timestamp like cbom-example does